+++ mc-all-platform-velocity/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.9.x (02/09/2025)
+- Update Geyser
+- Update EaglerXVelocity
+- Clean up Dockerfile
+- Fix some logs
+- Floodgate local linking
+
 ## 1.8.x (01/25/2025)
 - generate UUID and forwarding secret and save in config for persistance
 - configurable ICE servers to fix Eaglercraft voice chat